Env_dashboard/requirements.txt
streamlit>=1.55.0
streamlit-folium>=0.27.0
geemap>=0.37.0
earthengine-api>=0.1.349
geopandas>=1.1.0
shapely>=2.1.0
reportlab>=4.4.0
numpy>=1.24.0,<2.0
pandas>=2.3.0
matplotlib>=3.10.0
folium>=0.20.0
pillow>=12.0.0
requests>=2.33.0